+++ jest-tests.config.js
@@ -1,6 +1,6 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  testMatch: ['<rootDir>/src/**/*.spec.ts'],
+  testMatch: ['<rootDir>/src/**/*.spec.ts', '<rootDir>/src/**/*.test.ts'],
   coveragePathIgnorePatterns: ['\/__jest__\/'],
   moduleNameMapper: {
     "^@/(.*)": "<rootDir>/src/$1"