+++ test/src2/index.js
@@ -1,1 +1,1 @@
-console.log(JSON.stringify(require('../src1/styles.css')));
+console.log(JSON.stringify(require('../src1/styles.css'), null, 2));